Update description to give hints about stacking possibilities
It is now expected that several implementations of scope() may be used together
to combine a set of features, for example: define in private scope and run only
when all dependencies are ready.

Besides, the scope-bootstrap is now intended to be used for the definition of
other implementations of scope() in the global context.