Only use 'this' for globalContext in scope-bootstrap
this.global was used were available, to account for RingoJS:
* this is specific to each module
* the global object is found in this.global instead

I would rather avoid the hack though, as this.global is also defined
in Rhino and it is not the global object; in Rhino, it is a function,
and not an object, which could be checked using typeof.

However, since there is no standard that defines the type and behavior
of this.global, and there is no well agreed definition of this property,
this.global is not a reliable means to access the global scope.

I will rather define a specific version of scope-bootstrap for RingoJS,
and keep the default scope-bootstrap compatible with most implementations.