Add scope-bootstrap-ringojs for RingoJS
This is an alternative implementation of scope-bootstrap for RingoJS.
RingoJS does not give access to the global object in 'this', but makes
it available in 'this.global' instead.

Since this.global is also defined in other platforms, e.g. as a function
in Rhino, I preferred to keep the code to handle the specific case of RingoJS
outside the default scope-bootstrap.