Do not import properties from parent context
I expected this feature to make it easier to work with global
objects copied in the shared context, but it turned out wrong:
in the browsers, there are unexpected global properties which
may have the same name as a module, e.g. "home" which is a
function to replace the current page with the home page configured
by the user. In case such a property is found, the dependency
is never fetched e.g. when using AMD to load it asynchronously
when needed.

Due to the unexpected result, this feature will be deprecated.
I moved it to a separate module, scope-level2-shared-imports.js,
which will be deleted in a future version.