Updated ROADMAP
I now plan to use a simple, cross-platform system to run tests,
instead of using a different library for each platform (i.e. QUnit
for the browser, Jasmine for Node.js, ...)

As a consequence, I updated the ROADMAP to move examples and tests
running in Node.js to the next step.