Add examples for Node.js
I moved the modules to the folder 'node_modules' to facilitate dynamic
loading in Node.js.

The examples for Node.js are straightforward adaptations of the examples
for the browser, up to the last example, Example 5, which takes a specific
turn: dynamic loading with CommonJS require() in Node.js as opposed to
loading modules in any order (to prepare dynamic loading) in the browser.