Delete deprecated scope-level3-shared-imports
The script had been deprecated in v0.4.1, because it introduced a risk of
conflict between unexpected properties available globally in a specific
environment, e.g. 'home', a function related to the browser history, and
user-defined modules, e.g. 'home', a module related to the Home page of a
Web application.