Export modules for require() in scope-level3-commonjs
Instead of setting the module in 'exports', which does not refer
to the module currently loading, I now set the module in the object
returned by require(), which is the expected exports (and not read-only,
at least in Node.js). This allows for direct access to the module in
the CommonJS way, i.e. within a property of the same name on the object
returned by require().